Revert "Implement Live API for Samsung Now Bar"
