+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Fri Dec 05 06:56:30 UTC 2025
-gradle.version=8.8
+#Tue Dec 02 21:01:58 IST 2025
+gradle.version=8.2
