+++ app/src/main/AndroidManifest.xml
@@ -4,9 +4,6 @@
 
     <uses-permission android:name="android.permission.READ_CALENDAR" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" android:maxSdkVersion="32" />
-    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
-    <uses-permission android:name="android.permission.SCHEDULE_EXACT_ALARM" />
-    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
 
     <application
         android:allowBackup="true"
@@ -48,23 +45,6 @@
             <meta-data
                 android:name="android.appwidget.provider"
                 android:resource="@xml/calendar_widget_info" />
-        </receiver>
-
-        <receiver
-            android:name="com.antigravity.transparentcalendar.NotificationReceiver"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="com.antigravity.transparentcalendar.UPDATE_NOTIFICATION" />
-            </intent-filter>
-        </receiver>
-
-        <receiver
-            android:name="com.antigravity.transparentcalendar.BootReceiver"
-            android:exported="true"
-            android:enabled="true">
-            <intent-filter>
-                <action android:name="android.intent.action.BOOT_COMPLETED" />
-            </intent-filter>
         </receiver>
 
         <service
