+++ app/src/main/java/com/antigravity/transparentcalendar/CalendarWidgetFactory.kt
@@ -3,8 +3,11 @@ package com.antigravity.transparentcalendar
 import android.content.ContentUris
 import android.content.Context
 import android.content.Intent
+import android.database.Cursor
 import android.graphics.Color
 import android.provider.CalendarContract
+import android.text.format.DateUtils
+import android.view.View
 import android.widget.RemoteViews
 import android.widget.RemoteViewsService
 import java.text.SimpleDateFormat
@@ -47,7 +50,7 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
         if (position >= widgetItems.size) return RemoteViews(context.packageName, R.layout.widget_event_item)
 
         val item = widgetItems[position]
-
+        
         return when (item) {
             is WidgetListItem.HeaderItem -> {
                 val views = RemoteViews(context.packageName, R.layout.widget_date_header)
@@ -73,8 +76,8 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
                 if (color == 0) {
                     color = context.getColor(R.color.purple_500) // Default fallback
                 }
-
-
+                
+                
                 views.setInt(R.id.item_background, "setColorFilter", color)
 
                 val fillInIntent = Intent()
@@ -114,14 +117,14 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
 
         val headerFormat = SimpleDateFormat("EEE, d MMM", Locale.getDefault())
         val todayFormat = SimpleDateFormat("EEE, d MMM", Locale.getDefault())
-
+        
         val calendar = Calendar.getInstance()
         // Reset to start of today
         calendar.set(Calendar.HOUR_OF_DAY, 0)
         calendar.set(Calendar.MINUTE, 0)
         calendar.set(Calendar.SECOND, 0)
         calendar.set(Calendar.MILLISECOND, 0)
-
+        
         val todayStart = calendar.timeInMillis
         val todayDateText = todayFormat.format(Date(todayStart))
 
@@ -133,12 +136,12 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
             val dayStart = calendar.timeInMillis
             val currentDayYear = calendar.get(Calendar.YEAR)
             val currentDayOfYear = calendar.get(Calendar.DAY_OF_YEAR)
-
+            
             calendar.add(Calendar.DAY_OF_YEAR, 1)
             val dayEnd = calendar.timeInMillis
-
+            
             val dayDateText = headerFormat.format(Date(dayStart))
-
+            
             // Filter events that overlap with this day
             val dayEvents = events.filter { event ->
                 if (event.isAllDay) {
@@ -148,28 +151,28 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
                     val startDay = utcCalendar.get(Calendar.DAY_OF_YEAR)
                     // Simple comparable value: Year * 400 + Day (sufficient for near-term)
                     val startVal = startYear * 400 + startDay
-
+                    
                     utcCalendar.timeInMillis = event.end
                     val endYear = utcCalendar.get(Calendar.YEAR)
                     val endDay = utcCalendar.get(Calendar.DAY_OF_YEAR)
                     val endVal = endYear * 400 + endDay
-
+                    
                     val loopVal = currentDayYear * 400 + currentDayOfYear
-
+                    
                     loopVal >= startVal && loopVal < endVal
                 } else {
                     // Standard physical overlap for regular events
                     event.start < dayEnd && event.end > dayStart
                 }
             }
-
+            
             if (dayEvents.isNotEmpty()) {
                 // Add header if not today
                 if (dayDateText != todayDateText) {
                     val headerId = ("header_" + dayDateText).hashCode().toLong()
                     items.add(WidgetListItem.HeaderItem(dayDateText, headerId))
                 }
-
+                
                 for (event in dayEvents) {
                     // Create a unique ID for this specific day's entry of the event
                     val uniqueId = (event.id.toString() + "_" + dayStart).hashCode().toLong()
@@ -177,11 +180,86 @@ class CalendarWidgetFactory(private val context: Context, intent: Intent) : Remo
                 }
             }
         }
-
+        
         return items
     }
 
     private fun fetchEvents(): List<EventModel> {
-        return CalendarRepository(context).fetchEvents(7)
+        Log.d(TAG, "fetchEvents called")
+        val events = ArrayList<EventModel>()
+        
+        if (androidx.core.content.ContextCompat.checkSelfPermission(context, android.Manifest.permission.READ_CALENDAR) 
+            != android.content.pm.PackageManager.PERMISSION_GRANTED) {
+            Log.e(TAG, "READ_CALENDAR permission not granted")
+            return events
+        }
+
+
+
+        val projection = arrayOf(
+            CalendarContract.Instances.EVENT_ID,
+            CalendarContract.Instances.TITLE,
+            CalendarContract.Instances.BEGIN,
+            CalendarContract.Instances.END,
+            CalendarContract.Instances.DISPLAY_COLOR,
+            CalendarContract.Instances.ALL_DAY
+        )
+
+        val now = System.currentTimeMillis()
+        val endRange = now + DateUtils.WEEK_IN_MILLIS // Next 7 days
+
+        val selection = "${CalendarContract.Instances.END} >= ? AND ${CalendarContract.Instances.BEGIN} <= ?"
+        val selectionArgs = arrayOf(now.toString(), endRange.toString())
+
+        // Construct the URI for the instance table
+        val builder = CalendarContract.Instances.CONTENT_URI.buildUpon()
+        ContentUris.appendId(builder, now)
+        ContentUris.appendId(builder, endRange)
+
+        try {
+            val cursor: Cursor? = context.contentResolver.query(
+                builder.build(),
+                projection,
+                selection,
+                selectionArgs,
+                "${CalendarContract.Instances.BEGIN} ASC"
+            )
+            
+            if (cursor == null) {
+                Log.e(TAG, "Cursor is null")
+                return events
+            }
+
+            cursor.use {
+                val idIndex = it.getColumnIndex(CalendarContract.Instances.EVENT_ID)
+                val titleIndex = it.getColumnIndex(CalendarContract.Instances.TITLE)
+                val beginIndex = it.getColumnIndex(CalendarContract.Instances.BEGIN)
+                val endIndex = it.getColumnIndex(CalendarContract.Instances.END)
+                val colorIndex = it.getColumnIndex(CalendarContract.Instances.DISPLAY_COLOR)
+                val allDayIndex = it.getColumnIndex(CalendarContract.Instances.ALL_DAY)
+
+                if (idIndex == -1 || titleIndex == -1 || beginIndex == -1 || endIndex == -1) {
+                    Log.e(TAG, "Missing columns in cursor")
+                    return events
+                }
+
+                while (it.moveToNext()) {
+                    val id = it.getLong(idIndex)
+                    val title = it.getString(titleIndex) ?: "No Title"
+                    val start = it.getLong(beginIndex)
+                    val end = it.getLong(endIndex)
+                    val color = if (colorIndex != -1) it.getInt(colorIndex) else 0
+                    val isAllDay = if (allDayIndex != -1) it.getInt(allDayIndex) == 1 else false
+
+                    events.add(EventModel(id, title, start, end, color, isAllDay))
+                }
+            }
+            Log.d(TAG, "Fetched ${events.size} events")
+        } catch (e: Exception) {
+            Log.e(TAG, "Error fetching events", e)
+            e.printStackTrace()
+        }
+        
+        return events
     }
 }
