+++ app/src/main/java/com/antigravity/transparentcalendar/MainActivity.kt
@@ -11,23 +11,12 @@ import androidx.core.content.ContextCompat
 
 class MainActivity : AppCompatActivity() {
 
-    private val requestMultiplePermissionsLauncher =
-        registerForActivityResult(ActivityResultContracts.RequestMultiplePermissions()) { permissions ->
-            val calendarGranted = permissions[Manifest.permission.READ_CALENDAR] ?: false
-            val notificationsGranted = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                permissions[Manifest.permission.POST_NOTIFICATIONS] ?: false
+    private val requestPermissionLauncher =
+        registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted: Boolean ->
+            if (isGranted) {
+                Toast.makeText(this, "Permission granted! You can now add the widget.", Toast.LENGTH_LONG).show()
             } else {
-                true // Notification permission implicitly granted below Tiramisu
-            }
-
-            if (calendarGranted) {
-                Toast.makeText(this, "Permissions granted! You can now add the widget.", Toast.LENGTH_LONG).show()
-                // Refresh notifications if permissions granted
-                Thread {
-                    LiveNotificationManager.refreshNotifications(this)
-                }.start()
-            } else {
-                Toast.makeText(this, "Calendar permission denied. The widget will not show events.", Toast.LENGTH_LONG).show()
+                Toast.makeText(this, "Permission denied. The widget will not show events.", Toast.LENGTH_LONG).show()
             }
         }
 
@@ -35,37 +24,16 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        checkPermissionsAndRequest()
-
-        // Try to refresh notifications on launch
-        Thread {
-            LiveNotificationManager.refreshNotifications(this)
-        }.start()
+        checkPermissionAndRequest()
     }
 
-    private fun checkPermissionsAndRequest() {
-        val permissionsToRequest = mutableListOf<String>()
-
+    private fun checkPermissionAndRequest() {
         if (ContextCompat.checkSelfPermission(
                 this,
                 Manifest.permission.READ_CALENDAR
             ) != PackageManager.PERMISSION_GRANTED
         ) {
-            permissionsToRequest.add(Manifest.permission.READ_CALENDAR)
-        }
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            if (ContextCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.POST_NOTIFICATIONS
-                ) != PackageManager.PERMISSION_GRANTED
-            ) {
-                permissionsToRequest.add(Manifest.permission.POST_NOTIFICATIONS)
-            }
-        }
-
-        if (permissionsToRequest.isNotEmpty()) {
-            requestMultiplePermissionsLauncher.launch(permissionsToRequest.toTypedArray())
+            requestPermissionLauncher.launch(Manifest.permission.READ_CALENDAR)
         }
     }
 }
