+++ local.properties
@@ -1,0 +1,8 @@
+## This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+#
+# Location of the SDK. This is only used by Gradle.
+# For customization when using a Version Control System, please read the
+# header note.
+#Tue Dec 02 20:58:46 IST 2025
+sdk.dir=C\:\\Users\\singh\\AppData\\Local\\Android\\Sdk
